+++ Assets/Plugins/Appboy/Models/CardCategory.cs
@@ -1,9 +1,0 @@
-namespace Appboy.Models {
-  public enum CardCategory {
-    ADVERTISING,
-    ANNOUNCEMENTS,
-    NEWS,
-    SOCIAL,
-    NO_CATEGORY
-  }
-}