+++ Assets/Plugins/Appboy/Models/ClickAction.cs
@@ -1,6 +1,5 @@
 namespace Appboy.Models {
   public enum ClickAction {
-    NEWS_FEED,
     URI,
     NONE
   }